+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - More configurable default theme.
 - Option to require explicit confirmation of video/audio encoding by user
   before submitting jobs to Zencoder.
+- Preview draft revision via admin.
 
 - Bug fix: Improve video playback support on iOS and Android.
 - Bug fix: Missing translations for attribute/model names in admin.